refactor: voterId → src/lib/ 이동 (훅이 아닌 유틸)
