+++ src/hooks/usePick.ts
@@ -2,7 +2,7 @@
 
 import {useCallback, useEffect, useState} from 'react';
 
-import {getVoterId} from '@/hooks/voterId';
+import {getVoterId} from '@/lib/voterId';
 import {PickResult, PickType} from '@/types/vote';
 
 const DEFAULT_COUNTS: PickResult['counts'] = {COURSE_1: 0, COURSE_2: 0, TAKE_OUT: 0, pass: 0};
